backend/requirements.txt
asgiref==3.9.1
Django==6.0.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.6.0
django-jazzmin==3.0.1
psycopg2-binary==2.9.10
Pillow==11.1.0
requests==2.32.3
channels==4.2.0
daphne==4.2.1
redis==5.2.1
celery==5.4.0
django-redis==5.4.0
gunicorn==23.0.0
whitenoise==6.8.2
firebase-admin==6.5.0